Fix 'enter' not working in chat suggest widget Actually, we do need the EditorActions that I removed in #199099
