+++ src/vs/workbench/contrib/chat/browser/actions/chatExecuteActions.ts
@@ -4,16 +4,12 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { Codicon } from 'vs/base/common/codicons';
-import { KeyCode, KeyMod } from 'vs/base/common/keyCodes';
 import { ServicesAccessor } from 'vs/editor/browser/editorExtensions';
-import { localize, localize2 } from 'vs/nls';
+import { localize } from 'vs/nls';
 import { Action2, MenuId, registerAction2 } from 'vs/platform/actions/common/actions';
-import { KeybindingWeight } from 'vs/platform/keybinding/common/keybindingsRegistry';
 import { CHAT_CATEGORY } from 'vs/workbench/contrib/chat/browser/actions/chatActions';
 import { IChatWidget, IChatWidgetService } from 'vs/workbench/contrib/chat/browser/chat';
-import { IChatAgentService } from 'vs/workbench/contrib/chat/common/chatAgents';
-import { CONTEXT_CHAT_INPUT_HAS_TEXT, CONTEXT_CHAT_REQUEST_IN_PROGRESS, CONTEXT_IN_CHAT_INPUT } from 'vs/workbench/contrib/chat/common/chatContextKeys';
-import { chatAgentLeader } from 'vs/workbench/contrib/chat/common/chatParserTypes';
+import { CONTEXT_CHAT_INPUT_HAS_TEXT, CONTEXT_CHAT_REQUEST_IN_PROGRESS } from 'vs/workbench/contrib/chat/common/chatContextKeys';
 import { IChatService } from 'vs/workbench/contrib/chat/common/chatService';
 
 export interface IChatExecuteActionContext {
@@ -40,11 +36,6 @@ export class SubmitAction extends Action2 {
 				when: CONTEXT_CHAT_REQUEST_IN_PROGRESS.negate(),
 				group: 'navigation',
 			},
-			keybinding: {
-				when: CONTEXT_IN_CHAT_INPUT,
-				primary: KeyCode.Enter,
-				weight: KeybindingWeight.WorkbenchContrib
-			},
 		});
 	}
 
@@ -57,43 +48,8 @@ export class SubmitAction extends Action2 {
 	}
 }
 
-export class SubmitSecondaryAgentAction extends Action2 {
-	static readonly ID = 'workbench.action.chat.submitSecondaryAgent';
-
-	constructor() {
-		super({
-			id: SubmitSecondaryAgentAction.ID,
-			title: localize2('chat.label.submitSecondaryAgent', "Submit to Secondary Agent"),
-			f1: false,
-			category: CHAT_CATEGORY,
-			precondition: CONTEXT_CHAT_INPUT_HAS_TEXT,
-			keybinding: {
-				when: CONTEXT_IN_CHAT_INPUT,
-				primary: KeyMod.CtrlCmd | KeyCode.Enter,
-				weight: KeybindingWeight.WorkbenchContrib
-			},
-		});
-	}
-
-	run(accessor: ServicesAccessor, ...args: any[]) {
-		const context: IChatExecuteActionContext | undefined = args[0];
-
-		const agentService = accessor.get(IChatAgentService);
-		const secondaryAgent = agentService.getSecondaryAgent();
-		if (!secondaryAgent) {
-			return;
-		}
-
-		const widgetService = accessor.get(IChatWidgetService);
-		const widget = context?.widget ?? widgetService.lastFocusedWidget;
-		widget?.acceptInputWithPrefix(`${chatAgentLeader}${secondaryAgent.id}`);
-	}
-}
-
 export function registerChatExecuteActions() {
 	registerAction2(SubmitAction);
-	registerAction2(SubmitSecondaryAgentAction);
-
 	registerAction2(class CancelAction extends Action2 {
 		constructor() {
 			super({
